Restore BOARD_LETTERS constant and update .gitignore
Add *.vscode to .gitignore to enable the testing extension in vscode without pushing my personal settings to the repo.

This is also how I realized removing the BOARD_LETTERS constant and *forgetting* to add it as a class constant for the Board class. I'm keeping it as a module constant for now so other classes declared in board.py have access to it as needed.